Add popovers to category detail scores
